Allow email deliveries to update jobs
In order to verify that certain scheduled emails are delivered within the proper
window and only once, treating scheduled emails like cron jobs. Delivered emails
will go through SendGrid's parse API and then will be submitted to `/deadman`,
where we'll use the from address from the email to determine which job was
triggered.

*Important*: Handle downcased job IDs in email addresses

It's good practice to standardize email addresses as lower-case, but Cronut
wants job IDs to be upcased. Here we upcase the string before searching for
jobs.